requirements.txt
asgiref==3.8.1
chardet==5.2.0
comm==0.2.2
debugpy==1.8.2
Django==5.0.6
djangorestframework==3.15.2
executing==2.0.1
reportlab==4.2.2
sqlparse==0.5.0
typing_extensions==4.12.2
wcwidth==0.2.13
